Bug fix (?): Use the query row's key instead of the documentID to drive the data source.
IMPORTANT: This is a critical change, which fixes a problem when driving a table view from a view that generates multiple rows per document, as in the current example project -> 0..N units.

It should support all previous use cases where there was a 1-1 relationship between table rows and documents in the query. The important requirement is now that the key of each query row *must* be unique.